Remove "Command"/"Query" naming convention from CQRS
