--- src/Altinn.Broker.Application/ConfirmDownloadCommand/ConfirmDownloadCommandRequest.cs
@@ -1,10 +1,0 @@
-﻿
-using Altinn.Broker.Core.Domain;
-
-namespace Altinn.Broker.Application.ConfirmDownloadCommand;
-public class ConfirmDownloadCommandRequest
-{
-    public required CallerIdentity Token { get; set; }
-    public Guid FileTransferId { get; set; }
-    public bool IsLegacy { get; set; }
-}
