+++ src/Altinn.Broker.Application/DownloadFile/DownloadFileResponse.cs
@@ -1,0 +1,6 @@
+namespace Altinn.Broker.Application.DownloadFile;
+public class DownloadFileResponse
+{
+    public required string FileName { get; set; }
+    public required Stream DownloadStream { get; set; }
+}
