--- src/Altinn.Broker.Application/DownloadFileQuery/DownloadFileQueryHandler.cs
@@ -1,73 +1,0 @@
-using Altinn.Broker.Core.Application;
-using Altinn.Broker.Core.Domain.Enums;
-using Altinn.Broker.Core.Repositories;
-
-using Microsoft.Extensions.Logging;
-
-using OneOf;
-
-namespace Altinn.Broker.Application.DownloadFileQuery;
-public class DownloadFileQueryHandler : IHandler<DownloadFileQueryRequest, DownloadFileQueryResponse>
-{
-    private readonly IResourceRepository _resourceRepository;
-    private readonly IServiceOwnerRepository _serviceOwnerRepository;
-    private readonly IAuthorizationService _resourceRightsRepository;
-    private readonly IFileTransferRepository _fileTransferRepository;
-    private readonly IActorFileTransferStatusRepository _actorFileTransferStatusRepository;
-    private readonly IBrokerStorageService _brokerStorageService;
-    private readonly ILogger<DownloadFileQueryHandler> _logger;
-
-    public DownloadFileQueryHandler(IResourceRepository resourceRepository, IServiceOwnerRepository serviceOwnerRepository, IAuthorizationService resourceRightsRepository, IFileTransferRepository fileTransferRepository, IActorFileTransferStatusRepository actorFileTransferStatusRepository, IBrokerStorageService brokerStorageService, ILogger<DownloadFileQueryHandler> logger)
-    {
-        _resourceRepository = resourceRepository;
-        _serviceOwnerRepository = serviceOwnerRepository;
-        _resourceRightsRepository = resourceRightsRepository;
-        _fileTransferRepository = fileTransferRepository;
-        _actorFileTransferStatusRepository = actorFileTransferStatusRepository;
-        _brokerStorageService = brokerStorageService;
-        _logger = logger;
-    }
-
-    public async Task<OneOf<DownloadFileQueryResponse, Error>> Process(DownloadFileQueryRequest request, CancellationToken cancellationToken)
-    {
-        var fileTransfer = await _fileTransferRepository.GetFileTransfer(request.FileTransferId, cancellationToken);
-        if (fileTransfer is null)
-        {
-            return Errors.FileTransferNotFound;
-        }
-        if (fileTransfer.FileTransferStatusEntity.Status != FileTransferStatus.Published)
-        {
-            return Errors.FileTransferNotAvailable;
-        }
-        if (!fileTransfer.RecipientCurrentStatuses.Any(actorEvent => actorEvent.Actor.ActorExternalId == request.Token.Consumer))
-        {
-            return Errors.FileTransferNotFound;
-        }
-        if (string.IsNullOrWhiteSpace(fileTransfer?.FileLocation))
-        {
-            return Errors.NoFileUploaded;
-        }
-        var hasAccess = await _resourceRightsRepository.CheckUserAccess(fileTransfer.ResourceId, new List<ResourceAccessLevel> { ResourceAccessLevel.Read }, request.IsLegacy, cancellationToken);
-        if (!hasAccess)
-        {
-            return Errors.NoAccessToResource;
-        };
-        var resource = await _resourceRepository.GetResource(fileTransfer.ResourceId, cancellationToken);
-        if (resource is null)
-        {
-            return Errors.ResourceNotConfigured;
-        };
-        var serviceOwner = await _serviceOwnerRepository.GetServiceOwner(resource.ServiceOwnerId);
-        if (serviceOwner is null)
-        {
-            return Errors.ServiceOwnerNotConfigured;
-        };
-        var downloadStream = await _brokerStorageService.DownloadFile(serviceOwner, fileTransfer, cancellationToken);
-        await _actorFileTransferStatusRepository.InsertActorFileTransferStatus(request.FileTransferId, ActorFileTransferStatus.DownloadStarted, request.Token.Consumer, cancellationToken);
-        return new DownloadFileQueryResponse()
-        {
-            FileName = fileTransfer.FileName,
-            DownloadStream = downloadStream
-        };
-    }
-}
