--- src/Altinn.Broker.Application/DownloadFileQuery/DownloadFileQueryResponse.cs
@@ -1,6 +1,0 @@
-namespace Altinn.Broker.Application.DownloadFileQuery;
-public class DownloadFileQueryResponse
-{
-    public required string FileName { get; set; }
-    public required Stream DownloadStream { get; set; }
-}
