+++ src/Altinn.Broker.Application/GetFileTransferDetails/GetFileTransferDetailsRequest.cs
@@ -1,0 +1,10 @@
+
+using Altinn.Broker.Core.Domain;
+
+namespace Altinn.Broker.Application.GetFileTransferDetails;
+
+public class GetFileTransferDetailsRequest
+{
+    public required CallerIdentity Token { get; set; }
+    public Guid FileTransferId { get; set; }
+}
