+++ src/Altinn.Broker.Application/GetFileTransferOverview/GetFileTransferOverviewResponse.cs
@@ -1,0 +1,8 @@
+using Altinn.Broker.Core.Domain;
+
+namespace Altinn.Broker.Application.GetFileTransferOverview;
+
+public class GetFileTransferOverviewResponse
+{
+    public required FileTransferEntity FileTransfer { get; set; }
+}
