--- src/Altinn.Broker.Application/GetFileTransferOverviewQuery/GetFileTransferOverviewQueryResponse.cs
@@ -1,8 +1,0 @@
-using Altinn.Broker.Core.Domain;
-
-namespace Altinn.Broker.Application.GetFileTransferOverviewQuery;
-
-public class GetFileTransferOverviewQueryResponse
-{
-    public required FileTransferEntity FileTransfer { get; set; }
-}
