--- src/Altinn.Broker.Application/InitializeFileTransferCommand/InitializeFileTransferCommandRequest.cs
@@ -1,16 +1,0 @@
-﻿
-using Altinn.Broker.Core.Domain;
-
-namespace Altinn.Broker.Application.InitializeFileTransferCommand;
-public class InitializeFileTransferCommandRequest
-{
-    public required CallerIdentity Token { get; set; }
-    public required string ResourceId { get; set; }
-    public required string FileName { get; set; }
-    public required string SendersFileTransferReference { get; set; }
-    public required string SenderExternalId { get; set; }
-    public required List<string> RecipientExternalIds { get; set; }
-    public required Dictionary<string, string> PropertyList { get; set; }
-    public string? Checksum { get; set; }
-    public bool IsLegacy { get; set; }
-}
