+++ src/Altinn.Broker.Application/MalwareScanResults/MalwareScanResultHandler.cs
@@ -1,7 +1,7 @@
 
 using System.Text.Json;
 
-using Altinn.Broker.Application.ExpireFileTransferCommand;
+using Altinn.Broker.Application.ExpireFileTransfer;
 using Altinn.Broker.Core.Application;
 using Altinn.Broker.Core.Domain;
 using Altinn.Broker.Core.Repositories;
@@ -67,7 +67,7 @@ public class MalwareScanningResultHandler : IHandler<ScanResultData, Task>
         _logger.LogWarning("Suspicious scan result for file transfer {fileTransferId} with body {body}", fileTransferId, JsonSerializer.Serialize(data));
         await _fileTransferStatusRepository.InsertFileTransferStatus(fileTransferId, Core.Domain.Enums.FileTransferStatus.Failed, $"Malware scan failed: {data.ScanResultType}. Extra details: " + JsonSerializer.Serialize(data.ScanResultDetails), cancellationToken);
         await _eventBus.Publish(AltinnEventType.UploadFailed, fileTransfer.ResourceId, fileTransferIdFromUri, fileTransfer.Sender.ActorExternalId, cancellationToken);
-        _backgroundJobClient.Enqueue<ExpireFileTransferCommandHandler>(handler => handler.Process(new ExpireFileTransferCommandRequest
+        _backgroundJobClient.Enqueue<ExpireFileTransferHandler>(handler => handler.Process(new ExpireFileTransferRequest
         {
             FileTransferId = fileTransfer.FileTransferId,
             Force = true,
