+++ src/Altinn.Broker.Application/UploadFile/UploadFileRequest.cs
@@ -1,0 +1,13 @@
+
+using Altinn.Broker.Core.Domain;
+
+namespace Altinn.Broker.Application.UploadFile;
+
+public class UploadFileRequest
+{
+    public Guid FileTransferId { get; set; }
+    public required CallerIdentity Token { get; set; }
+    public required Stream UploadStream { get; set; }
+    public bool IsLegacy { get; set; }
+    public long ContentLength { get; set; }
+}
