+++ tests/Altinn.Broker.Tests/LegacyFileControllerTests.cs
@@ -10,6 +10,7 @@ using Altinn.Broker.Enums;
 using Altinn.Broker.Models;
 using Altinn.Broker.Tests.Factories;
 using Altinn.Broker.Tests.Helpers;
+
 using Xunit;
 
 namespace Altinn.Broker.Tests;
